==> notebooks/6_analysis_of_cohort_behavior.ipynb <==
import numpy as np
import pandas as pd
import plotly.express as px
import plotly.graph_objects as go
from plotly.subplots import make_subplots
import re

import seaborn as sns
import matplotlib.pyplot as plt
sns.set_style("white")
%matplotlib inline

yearly=pd.read_pickle('yearly.pkl')

for col in yearly.columns:
    print(col)

yearly.head()

yearly.groupby('first_gift_year')

y15=yearly.query('first_gift_year == "2015-01-01"')

y15.id.count()

y15.behavior_2017.value_counts()

y15.query('behavior_2017 == "lapse_repeat"')

yearly2=yearly[['id',
'first_gift_year',
'years_donated',
'perc_years_donated',
'gift_pattern',
'donor_profile',
'behavior_2013',
'behavior_2014',
'behavior_2015',
'behavior_2016',
'behavior_2017',
'gainloss_2013',
'gainloss_2014',
'gainloss_2015',
'gainloss_2016',
'gainloss_2017',
'gainloss_2014_amt',
'gainloss_2015_amt',
'gainloss_2016_amt',
'gainloss_2017_amt']].copy()

y13=yearly2.query('first_gift_year=="2013-01-01"')

y13

print(y16.behavior_2013.value_counts())
print(y16.behavior_2014.value_counts())
print(y16.behavior_2015.value_counts())
print(y16.behavior_2016.value_counts())
print(y16.behavior_2017.value_counts())

y13.query('behavior_2017 == "lapse_repeat"')

donations_13to17=pd.read_pickle('donations_13to17.pkl')

d13=donations_13to17.query('first_gift_year=="2013-01-01"')
d14=donations_13to17.query('first_gift_year=="2014-01-01"')
d15=donations_13to17.query('first_gift_year=="2015-01-01"')
d16=donations_13to17.query('first_gift_year=="2016-01-01"')
d17=donations_13to17.query('first_gift_year=="2017-01-01"')

print(d16.groupby('behavior_2013').count()['id'])
print(d16.groupby('behavior_2014').count()['id'])
print(d16.groupby('behavior_2015').count()['id'])
print(d16.groupby('behavior_2016').count()['id'])
print(d17.groupby('behavior_2017').count()['id'])
