==> notebooks/7_prepare_lifecycle_datasets.ipynb <==
import numpy as np
import pandas as pd
import plotly.express as px
import plotly.graph_objects as go
from plotly.subplots import make_subplots
import re

import seaborn as sns
import matplotlib.pyplot as plt
sns.set_style("white")
%matplotlib inline

#new dataset byphase
new_donations_byphase=pd.read_csv('../data/donations_by_lifecycle/donations_by_phase.csv')

new_donations_byphase.shape

# rename columns
new_donations_byphase.columns=[
'id',
'first_gift_year',
'lapsed_amount',
'recovered_amount',
'lapsed_count',
'recovered_count',
'max_consec_giving_years',
'total_giving_years',
'amount_2013',
'amount_2014',
'amount_2015',
'amount_2016',
'amount_2017',
'count_2013',
'count_2014',
'count_2015',
'count_2016',
'count_2017',
'lapsed_amount_2013',
'lapsed_amount_2014',
'lapsed_amount_2015',
'lapsed_amount_2016',
'lapsed_amount_2017',
'upgraded_amount_2013',
'upgraded_amount_2014',
'upgraded_amount_2015',
'upgraded_amount_2016',
'upgraded_amount_2017',
'downgraded_amount_2013',
'downgraded_amount_2014',
'downgraded_amount_2015',
'downgraded_amount_2016',
'downgraded_amount_2017',
'amount_yr0',
'amount_yr1',
'amount_yr2',
'amount_yr3',
'amount_yr4',
'count_yr0',
'count_yr1',
'count_yr2',
'count_yr3',
'count_yr4',
'lapsed_amount_yr0',
'lapsed_amount_yr1',
'lapsed_amount_yr2',
'lapsed_amount_yr3',
'lapsed_amount_yr4',
'upgraded_amount_yr0',
'upgraded_amount_yr1',
'upgraded_amount_yr2',
'upgraded_amount_yr3',
'upgraded_amount_yr4',
'downgraded_amount_yr0',
'downgraded_amount_yr1',
'downgraded_amount_yr2',
'downgraded_amount_yr3',
'downgraded_amount_yr4',
'amount_01_2013',
'amount_02_2013',
'amount_03_2013',
'amount_04_2013',
'amount_05_2013',
'amount_06_2013',
'amount_07_2013',
'amount_08_2013',
'amount_09_2013',
'amount_10_2013',
'amount_11_2013',
'amount_12_2013',
'amount_01_2014',
'amount_02_2014',
'amount_03_2014',
'amount_04_2014',
'amount_05_2014',
'amount_06_2014',
'amount_07_2014',
'amount_08_2014',
'amount_09_2014',
'amount_10_2014',
'amount_11_2014',
'amount_12_2014',
'amount_01_2015',
'amount_02_2015',
'amount_03_2015',
'amount_04_2015',
'amount_05_2015',
'amount_06_2015',
'amount_07_2015',
'amount_08_2015',
'amount_09_2015',
'amount_10_2015',
'amount_11_2015',
'amount_12_2015',
'amount_01_2016',
'amount_02_2016',
'amount_03_2016',
'amount_04_2016',
'amount_05_2016',
'amount_06_2016',
'amount_07_2016',
'amount_08_2016',
'amount_09_2016',
'amount_10_2016',
'amount_11_2016',
'amount_12_2016',
'amount_01_2017',
'amount_02_2017',
'amount_03_2017',
'amount_04_2017',
'amount_05_2017',
'amount_06_2017',
'amount_07_2017',
'amount_08_2017',
'amount_09_2017',
'amount_10_2017',
'amount_11_2017',
'amount_12_2017',
'initial',
'first_gift',
'second_year',
'second_with_gap',
'multi_year',
'key_multi_year',
'recovered_multi_year',
'inconsistent_multi_year ',
'recovered_key_multi_year',
'lapsed_after_first',
'lapsed_after_first_long',
'lapsed_after_second',
'lapsed_multi_year',
'lapsed_long_multi_year',
'lapsed_key_multi_year',
'lapsed_long',
]

#giving patterns
conditions = [(new_donations_byphase[('amount_2013')] >0) & (new_donations_byphase[('amount_2014')] ==0) & (new_donations_byphase[('amount_2015')]==0) & (new_donations_byphase[('amount_2016')]==0) & (new_donations_byphase[('amount_2017')]== 0) ,
(new_donations_byphase[('amount_2013')] ==0) & (new_donations_byphase[('amount_2014')] >0) & (new_donations_byphase[('amount_2015')]==0) & (new_donations_byphase[('amount_2016')]==0) & (new_donations_byphase[('amount_2017')]== 0) ,
(new_donations_byphase[('amount_2013')] ==0) & (new_donations_byphase[('amount_2014')] ==0) & (new_donations_byphase[('amount_2015')]>0) & (new_donations_byphase[('amount_2016')]==0) & (new_donations_byphase[('amount_2017')]== 0) ,
(new_donations_byphase[('amount_2013')] ==0) & (new_donations_byphase[('amount_2014')] ==0) & (new_donations_byphase[('amount_2015')]==0) & (new_donations_byphase[('amount_2016')]>0) & (new_donations_byphase[('amount_2017')]== 0) ,
(new_donations_byphase[('amount_2013')] ==0) & (new_donations_byphase[('amount_2014')] ==0) & (new_donations_byphase[('amount_2015')]==0) & (new_donations_byphase[('amount_2016')]==0) & (new_donations_byphase[('amount_2017')]> 0),
(new_donations_byphase[('amount_2013')] >0) & (new_donations_byphase[('amount_2014')] >0) & (new_donations_byphase[('amount_2015')]==0) & (new_donations_byphase[('amount_2016')]==0) & (new_donations_byphase[('amount_2017')]== 0) ,
(new_donations_byphase[('amount_2013')] >0) & (new_donations_byphase[('amount_2014')] ==0) & (new_donations_byphase[('amount_2015')]>0) & (new_donations_byphase[('amount_2016')]==0) & (new_donations_byphase[('amount_2017')]== 0) ,
(new_donations_byphase[('amount_2013')] >0) & (new_donations_byphase[('amount_2014')] ==0) & (new_donations_byphase[('amount_2015')]==0) & (new_donations_byphase[('amount_2016')]>0) & (new_donations_byphase[('amount_2017')]== 0) ,
(new_donations_byphase[('amount_2013')] >0) & (new_donations_byphase[('amount_2014')] ==0) & (new_donations_byphase[('amount_2015')]==0) & (new_donations_byphase[('amount_2016')]==0) & (new_donations_byphase[('amount_2017')]> 0) ,
(new_donations_byphase[('amount_2013')] ==0) & (new_donations_byphase[('amount_2014')] >0) & (new_donations_byphase[('amount_2015')]>0) & (new_donations_byphase[('amount_2016')]==0) & (new_donations_byphase[('amount_2017')]== 0) ,
(new_donations_byphase[('amount_2013')] ==0) & (new_donations_byphase[('amount_2014')] >0) & (new_donations_byphase[('amount_2015')]==0) & (new_donations_byphase[('amount_2016')]>0) & (new_donations_byphase[('amount_2017')]== 0) ,
(new_donations_byphase[('amount_2013')] ==0) & (new_donations_byphase[('amount_2014')] >0) & (new_donations_byphase[('amount_2015')]==0) & (new_donations_byphase[('amount_2016')]==0) & (new_donations_byphase[('amount_2017')]> 0) ,
(new_donations_byphase[('amount_2013')] ==0) & (new_donations_byphase[('amount_2014')] ==0) & (new_donations_byphase[('amount_2015')]>0) & (new_donations_byphase[('amount_2016')]>0) & (new_donations_byphase[('amount_2017')]== 0) ,
(new_donations_byphase[('amount_2013')] ==0) & (new_donations_byphase[('amount_2014')] ==0) & (new_donations_byphase[('amount_2015')]>0) & (new_donations_byphase[('amount_2016')]==0) & (new_donations_byphase[('amount_2017')]> 0) ,
(new_donations_byphase[('amount_2013')] ==0) & (new_donations_byphase[('amount_2014')] ==0) & (new_donations_byphase[('amount_2015')]==0) & (new_donations_byphase[('amount_2016')]>0) & (new_donations_byphase[('amount_2017')]> 0),
(new_donations_byphase[('amount_2013')] >0) & (new_donations_byphase[('amount_2014')] >0) & (new_donations_byphase[('amount_2015')]>0) & (new_donations_byphase[('amount_2016')]==0) & (new_donations_byphase[('amount_2017')]== 0) ,
(new_donations_byphase[('amount_2013')] >0) & (new_donations_byphase[('amount_2014')] >0) & (new_donations_byphase[('amount_2015')]==0) & (new_donations_byphase[('amount_2016')]>0) & (new_donations_byphase[('amount_2017')]== 0) ,
(new_donations_byphase[('amount_2013')] >0) & (new_donations_byphase[('amount_2014')] >0) & (new_donations_byphase[('amount_2015')]==0) & (new_donations_byphase[('amount_2016')]==0) & (new_donations_byphase[('amount_2017')]> 0) ,
(new_donations_byphase[('amount_2013')] >0) & (new_donations_byphase[('amount_2014')] ==0) & (new_donations_byphase[('amount_2015')]>0) & (new_donations_byphase[('amount_2016')]>0) & (new_donations_byphase[('amount_2017')]== 0) ,
(new_donations_byphase[('amount_2013')] >0) & (new_donations_byphase[('amount_2014')] ==0) & (new_donations_byphase[('amount_2015')]>0) & (new_donations_byphase[('amount_2016')]==0) & (new_donations_byphase[('amount_2017')]> 0) ,
(new_donations_byphase[('amount_2013')] >0) & (new_donations_byphase[('amount_2014')] ==0) & (new_donations_byphase[('amount_2015')]==0) & (new_donations_byphase[('amount_2016')]>0) & (new_donations_byphase[('amount_2017')]> 0) ,
(new_donations_byphase[('amount_2013')] ==0) & (new_donations_byphase[('amount_2014')] >0) & (new_donations_byphase[('amount_2015')]>0) & (new_donations_byphase[('amount_2016')]>0) & (new_donations_byphase[('amount_2017')]== 0) ,
(new_donations_byphase[('amount_2013')] ==0) & (new_donations_byphase[('amount_2014')] >0) & (new_donations_byphase[('amount_2015')]>0) & (new_donations_byphase[('amount_2016')]==0) & (new_donations_byphase[('amount_2017')]> 0) ,
(new_donations_byphase[('amount_2013')] ==0) & (new_donations_byphase[('amount_2014')] >0) & (new_donations_byphase[('amount_2015')]==0) & (new_donations_byphase[('amount_2016')]>0) & (new_donations_byphase[('amount_2017')]> 0) ,
(new_donations_byphase[('amount_2013')] ==0) & (new_donations_byphase[('amount_2014')] ==0) & (new_donations_byphase[('amount_2015')]>0) & (new_donations_byphase[('amount_2016')]>0) & (new_donations_byphase[('amount_2017')]> 0),
(new_donations_byphase[('amount_2013')] >0) & (new_donations_byphase[('amount_2014')] >0) & (new_donations_byphase[('amount_2015')]>0) & (new_donations_byphase[('amount_2016')]>0) & (new_donations_byphase[('amount_2017')]== 0) ,
(new_donations_byphase[('amount_2013')] >0) & (new_donations_byphase[('amount_2014')] >0) & (new_donations_byphase[('amount_2015')]>0) & (new_donations_byphase[('amount_2016')]==0) & (new_donations_byphase[('amount_2017')]> 0) ,
(new_donations_byphase[('amount_2013')] >0) & (new_donations_byphase[('amount_2014')] >0) & (new_donations_byphase[('amount_2015')]==0) & (new_donations_byphase[('amount_2016')]>0) & (new_donations_byphase[('amount_2017')]> 0) ,
(new_donations_byphase[('amount_2013')] >0) & (new_donations_byphase[('amount_2014')] ==0) & (new_donations_byphase[('amount_2015')]>0) & (new_donations_byphase[('amount_2016')]>0) & (new_donations_byphase[('amount_2017')]> 0) ,
(new_donations_byphase[('amount_2013')] ==0) & (new_donations_byphase[('amount_2014')] >0) & (new_donations_byphase[('amount_2015')]>0) & (new_donations_byphase[('amount_2016')]>0) & (new_donations_byphase[('amount_2017')]> 0),
(new_donations_byphase[('amount_2013')] >0) & (new_donations_byphase[('amount_2014')] >0) & (new_donations_byphase[('amount_2015')]>0) & (new_donations_byphase[('amount_2016')]>0) & (new_donations_byphase[('amount_2017')]> 0)
]

choices = ['1,', '2,', 
'3,', 
'4,', 
'5,', 
'1,2,',
'1,3,', 
'1,4,', 
'1,5,', 
'2,3,', 
'2,4,', 
'2,5,', 
'3,4,', 
'3,5,', 
'4,5,', 
'1,2,3,', 
'1,2,4,', 
'1,2,5,', 
'1,3,4,', 
'1,3,5,', 
'1,4,5,', 
'2,3,4,', 
'2,3,5,', 
'2,4,5,', 
'3,4,5,', 
'1,2,3,4,', 
'1,2,3,5,', 
'1,2,4,5,', 
'1,3,4,5,', 
'2,3,4,5,', 
'1,2,3,4,5,' 
]

new_donations_byphase.loc[:,'gift_pattern'] = np.select(conditions, choices, default='none')


#giving patterns
conditions=[(new_donations_byphase[('amount_yr0')] >0) & (new_donations_byphase[('amount_yr1')] ==0) & (new_donations_byphase[('amount_yr2')]==0) & (new_donations_byphase[('amount_yr3')]==0) & (new_donations_byphase[('amount_yr4')]== 0) ,
(new_donations_byphase[('amount_yr0')] >0) & (new_donations_byphase[('amount_yr1')] >0) & (new_donations_byphase[('amount_yr2')]==0) & (new_donations_byphase[('amount_yr3')]==0) & (new_donations_byphase[('amount_yr4')]== 0) ,
(new_donations_byphase[('amount_yr0')] >0) & (new_donations_byphase[('amount_yr1')] ==0) & (new_donations_byphase[('amount_yr2')]>0) & (new_donations_byphase[('amount_yr3')]==0) & (new_donations_byphase[('amount_yr4')]== 0) ,
(new_donations_byphase[('amount_yr0')] >0) & (new_donations_byphase[('amount_yr1')] ==0) & (new_donations_byphase[('amount_yr2')]==0) & (new_donations_byphase[('amount_yr3')]>0) & (new_donations_byphase[('amount_yr4')]== 0) ,
(new_donations_byphase[('amount_yr0')] >0) & (new_donations_byphase[('amount_yr1')] ==0) & (new_donations_byphase[('amount_yr2')]==0) & (new_donations_byphase[('amount_yr3')]==0) & (new_donations_byphase[('amount_yr4')]> 0) ,
(new_donations_byphase[('amount_yr0')] >0) & (new_donations_byphase[('amount_yr1')] >0) & (new_donations_byphase[('amount_yr2')]>0) & (new_donations_byphase[('amount_yr3')]==0) & (new_donations_byphase[('amount_yr4')]== 0) ,
(new_donations_byphase[('amount_yr0')] >0) & (new_donations_byphase[('amount_yr1')] >0) & (new_donations_byphase[('amount_yr2')]==0) & (new_donations_byphase[('amount_yr3')]>0) & (new_donations_byphase[('amount_yr4')]== 0) ,
(new_donations_byphase[('amount_yr0')] >0) & (new_donations_byphase[('amount_yr1')] >0) & (new_donations_byphase[('amount_yr2')]==0) & (new_donations_byphase[('amount_yr3')]==0) & (new_donations_byphase[('amount_yr4')]> 0) ,
(new_donations_byphase[('amount_yr0')] >0) & (new_donations_byphase[('amount_yr1')] ==0) & (new_donations_byphase[('amount_yr2')]>0) & (new_donations_byphase[('amount_yr3')]>0) & (new_donations_byphase[('amount_yr4')]== 0) ,
(new_donations_byphase[('amount_yr0')] >0) & (new_donations_byphase[('amount_yr1')] ==0) & (new_donations_byphase[('amount_yr2')]>0) & (new_donations_byphase[('amount_yr3')]==0) & (new_donations_byphase[('amount_yr4')]>0) ,
(new_donations_byphase[('amount_yr0')] >0) & (new_donations_byphase[('amount_yr1')] ==0) & (new_donations_byphase[('amount_yr2')]==0) & (new_donations_byphase[('amount_yr3')]>0) & (new_donations_byphase[('amount_yr4')]> 0) ,
(new_donations_byphase[('amount_yr0')] >0) & (new_donations_byphase[('amount_yr1')] >0) & (new_donations_byphase[('amount_yr2')]>0) & (new_donations_byphase[('amount_yr3')]>0) & (new_donations_byphase[('amount_yr4')]== 0) ,
(new_donations_byphase[('amount_yr0')] >0) & (new_donations_byphase[('amount_yr1')] >0) & (new_donations_byphase[('amount_yr2')]>0) & (new_donations_byphase[('amount_yr3')]==0) & (new_donations_byphase[('amount_yr4')]>0) ,
(new_donations_byphase[('amount_yr0')] >0) & (new_donations_byphase[('amount_yr1')] >0) & (new_donations_byphase[('amount_yr2')]==0) & (new_donations_byphase[('amount_yr3')]>0) & (new_donations_byphase[('amount_yr4')]>0) ,
(new_donations_byphase[('amount_yr0')] >0) & (new_donations_byphase[('amount_yr1')] ==0) & (new_donations_byphase[('amount_yr2')]>0) & (new_donations_byphase[('amount_yr3')]>0) & (new_donations_byphase[('amount_yr4')]>0) ,
(new_donations_byphase[('amount_yr0')] >0) & (new_donations_byphase[('amount_yr1')] >0) & (new_donations_byphase[('amount_yr2')]>0) & (new_donations_byphase[('amount_yr3')]>0) & (new_donations_byphase[('amount_yr4')]>0)]


choices = ['0,',
'0,1,',
'0,2,',
'0,3,',
'0,4,',
'0,1,2,',
'0,1,3,',
'0,1,4,',
'0,2,3,',
'0,2,4,',
'0,3,4',
'0,1,2,3,',
'0,1,2,4,',
'0,1,3,4,',
'0,2,3,4,',
'0,1,2,3,4,']

new_donations_byphase.loc[:,'cohort_gift_pattern'] = np.select(conditions, choices, default='none')

new_donations_byphase.columns.tolist()

cols=['cohort_gift_pattern',
 'amount_yr0',
 'amount_yr1',
 'amount_yr2',
 'amount_yr3',
 'amount_yr4',
 'amount_2013',
 'amount_2014',
 'amount_2015',
 'amount_2016',
 'amount_2017',
 'id',
 'gift_pattern',
 'first_gift_year',
 'lapsed_amount',
 'recovered_amount',
 'lapsed_count',
 'recovered_count',
 'max_consec_giving_years',
 'total_giving_years',
 'amount_2013',
 'amount_2014',
 'amount_2015',
 'amount_2016',
 'amount_2017',
 'count_2013',
 'count_2014',
 'count_2015',
 'count_2016',
 'count_2017',
 'lapsed_amount_2013',
 'lapsed_amount_2014',
 'lapsed_amount_2015',
 'lapsed_amount_2016',
 'lapsed_amount_2017',
 'upgraded_amount_2013',
 'upgraded_amount_2014',
 'upgraded_amount_2015',
 'upgraded_amount_2016',
 'upgraded_amount_2017',
 'downgraded_amount_2013',
 'downgraded_amount_2014',
 'downgraded_amount_2015',
 'downgraded_amount_2016',
 'downgraded_amount_2017',
 'amount_yr0',
 'amount_yr1',
 'amount_yr2',
 'amount_yr3',
 'amount_yr4',
 'count_yr0',
 'count_yr1',
 'count_yr2',
 'count_yr3',
 'count_yr4',
 'lapsed_amount_yr0',
 'lapsed_amount_yr1',
 'lapsed_amount_yr2',
 'lapsed_amount_yr3',
 'lapsed_amount_yr4',
 'upgraded_amount_yr0',
 'upgraded_amount_yr1',
 'upgraded_amount_yr2',
 'upgraded_amount_yr3',
 'upgraded_amount_yr4',
 'downgraded_amount_yr0',
 'downgraded_amount_yr1',
 'downgraded_amount_yr2',
 'downgraded_amount_yr3',
 'downgraded_amount_yr4',
 'amount_01_2013',
 'amount_02_2013',
 'amount_03_2013',
 'amount_04_2013',
 'amount_05_2013',
 'amount_06_2013',
 'amount_07_2013',
 'amount_08_2013',
 'amount_09_2013',
 'amount_10_2013',
 'amount_11_2013',
 'amount_12_2013',
 'amount_01_2014',
 'amount_02_2014',
 'amount_03_2014',
 'amount_04_2014',
 'amount_05_2014',
 'amount_06_2014',
 'amount_07_2014',
 'amount_08_2014',
 'amount_09_2014',
 'amount_10_2014',
 'amount_11_2014',
 'amount_12_2014',
 'amount_01_2015',
 'amount_02_2015',
 'amount_03_2015',
 'amount_04_2015',
 'amount_05_2015',
 'amount_06_2015',
 'amount_07_2015',
 'amount_08_2015',
 'amount_09_2015',
 'amount_10_2015',
 'amount_11_2015',
 'amount_12_2015',
 'amount_01_2016',
 'amount_02_2016',
 'amount_03_2016',
 'amount_04_2016',
 'amount_05_2016',
 'amount_06_2016',
 'amount_07_2016',
 'amount_08_2016',
 'amount_09_2016',
 'amount_10_2016',
 'amount_11_2016',
 'amount_12_2016',
 'amount_01_2017',
 'amount_02_2017',
 'amount_03_2017',
 'amount_04_2017',
 'amount_05_2017',
 'amount_06_2017',
 'amount_07_2017',
 'amount_08_2017',
 'amount_09_2017',
 'amount_10_2017',
 'amount_11_2017',
 'amount_12_2017',
 'initial',
 'first_gift',
 'second_year',
 'second_with_gap',
 'multi_year',
 'key_multi_year',
 'recovered_multi_year',
 'inconsistent_multi_year ',
 'recovered_key_multi_year',
 'lapsed_after_first',
 'lapsed_after_first_long',
 'lapsed_after_second',
 'lapsed_multi_year',
 'lapsed_long_multi_year',
 'lapsed_key_multi_year',
 'lapsed_long']

new_donations=new_donations_byphase[cols]

new_donations

new_donations.cohort_gift_pattern.value_counts()

new_donations.groupby(['cohort_gift_pattern','first_gift_year'])['id'].count()

new_donations.to_pickle('new_donations.pkl')

new_donations_sample=new_donations.sample(n=2000)

new_donations_sample.to_csv('new_donations_sample.csv')

for i in new_donations_byphase.columns:
    x = new_donations_byphase[i].value_counts()
    print("Column name is:",i,"and it value is:",x)
    print()

#keep new columns
new_columns=new_donations_byphase[[
'id',
'lifecycle_initial',
'lifecycle_first_gift',
'lifecycle_second_year',
'lifecycle_second_with_gap',
'lifecycle_multi_year',
'lifecycle_key_multi_year',
'lifecycle_recovered_multi_year',
'lifecycle_inconsistent_multi_year ',
'lifecycle_recovered_key_multi_year',
'lifecycle_lapsed_after_first',
'lifecycle_lapsed_after_first_long',
'lifecycle_lapsed_after_second',
'lifecycle_lapsed_multi_year',
'lifecycle_lapsed_long_multi_year',
'lifecycle_lapsed_key_multi_year',
'lifecycle_lapsed_long'
]].copy()

# merge new columns with old dataset
donations_lifecycle=donations_13to17.merge(new_columns, on="id")

df_tr = samp.transpose()

df_tr

with pd.option_context('display.max_rows', None, 'display.max_columns', 10):
    print(df_tr)
