==> notebooks/10_new_dataset.ipynb <==
import numpy as np
import pandas as pd
import plotly.express as px
import plotly.graph_objects as go
from plotly.subplots import make_subplots
#import re

import seaborn as sns
import matplotlib.pyplot as plt
sns.set_style("white")
%matplotlib inline

donations_cohort_with_cycle = pd.read_csv('../data/donations_cohort_with_cycle.csv')

donations_cohort_with_cycle_sample = donations_cohort_with_cycle.sample(n=100)

donations_cohort_with_cycle_sample.to_csv('donations_cohort_with_cycle_sample.csv')

# for columns in donations_cohort_with_cycle.columns:
#     print(columns)

# rename columns
donations_cohort_with_cycle.columns=[
'id',
'first_gift_year',
'lapsed_amount',
'recovered_amount',
'lapsed_count',
'recovered_count',
'max_consec_giving_years',
'total_giving_years',
'amount_2013',
'amount_2014',
'amount_2015',
'amount_2016',
'amount_2017',
'count_2013',
'count_2014',
'count_2015',
'count_2016',
'count_2017',
'lapsed_amount_2013',
'lapsed_amount_2014',
'lapsed_amount_2015',
'lapsed_amount_2016',
'lapsed_amount_2017',
'upgraded_amount_2013',
'upgraded_amount_2014',
'upgraded_amount_2015',
'upgraded_amount_2016',
'upgraded_amount_2017',
'downgraded_amount_2013',
'downgraded_amount_2014',
'downgraded_amount_2015',
'downgraded_amount_2016',
'downgraded_amount_2017',
'amount_yr0',
'amount_yr1',
'amount_yr2',
'amount_yr3',
'amount_yr4',
'count_yr0',
'count_yr1',
'count_yr2',
'count_yr3',
'count_yr4',
'lapsed_amount_yr0',
'lapsed_amount_yr1',
'lapsed_amount_yr2',
'lapsed_amount_yr3',
'lapsed_amount_yr4',
'upgraded_amount_yr0',
'upgraded_amount_yr1',
'upgraded_amount_yr2',
'upgraded_amount_yr3',
'upgraded_amount_yr4',
'downgraded_amount_yr0',
'downgraded_amount_yr1',
'downgraded_amount_yr2',
'downgraded_amount_yr3',
'downgraded_amount_yr4',
'amount_01_2013',
'amount_02_2013',
'amount_03_2013',
'amount_04_2013',
'amount_05_2013',
'amount_06_2013',
'amount_07_2013',
'amount_08_2013',
'amount_09_2013',
'amount_10_2013',
'amount_11_2013',
'amount_12_2013',
'amount_01_2014',
'amount_02_2014',
'amount_03_2014',
'amount_04_2014',
'amount_05_2014',
'amount_06_2014',
'amount_07_2014',
'amount_08_2014',
'amount_09_2014',
'amount_10_2014',
'amount_11_2014',
'amount_12_2014',
'amount_01_2015',
'amount_02_2015',
'amount_03_2015',
'amount_04_2015',
'amount_05_2015',
'amount_06_2015',
'amount_07_2015',
'amount_08_2015',
'amount_09_2015',
'amount_10_2015',
'amount_11_2015',
'amount_12_2015',
'amount_01_2016',
'amount_02_2016',
'amount_03_2016',
'amount_04_2016',
'amount_05_2016',
'amount_06_2016',
'amount_07_2016',
'amount_08_2016',
'amount_09_2016',
'amount_10_2016',
'amount_11_2016',
'amount_12_2016',
'amount_01_2017',
'amount_02_2017',
'amount_03_2017',
'amount_04_2017',
'amount_05_2017',
'amount_06_2017',
'amount_07_2017',
'amount_08_2017',
'amount_09_2017',
'amount_10_2017',
'amount_11_2017',
'amount_12_2017',
'initial',
'first_gift',
'second_year',
'second_with_gap',
'multi_year',
'key_multi_year',
'recovered_multi_year',
'inconsistent_multi_year ',
'recovered_key_multi_year',
'lapsed_after_first',
'lapsed_after_first_long',
'lapsed_after_second',
'lapsed_multi_year',
'lapsed_long_multi_year',
'lapsed_key_multi_year',
'lapsed_long',
'cohort_0_first_gift',
'cohort_1_lapsed_after_first',
'cohort_1_second_year',
'cohort_2_lapsed_after_first_long',
'cohort_2_lapsed_after_second',
'cohort_2_multi_year',
'cohort_2_second_with_gap',
'cohort_3_first_gift',
'cohort_3_inconsistent_multi_year',
'cohort_3_key_multi_year',
'cohort_3_lapsed_after_second',
'cohort_3_lapsed_long',
'cohort_3_lapsed_multi_year',
'cohort_4_first_gift',
'cohort_4_inconsistent_multi_year',
'cohort_4_key_multi_year',
'cohort_4_lapsed_after_first',
'cohort_4_lapsed_key_multi_year',
'cohort_4_lapsed_long',
'cohort_4_lapsed_long_multi_year',
'cohort_4_lapsed_multi_year',
'cohort_4_multi_year',
'cohort_4_recovered_multi_year',
'cohort_4_second_year',
'cohort_5_first_gift',
'cohort_5_inconsistent_multi_year',
'cohort_5_key_multi_year',
'cohort_5_lapsed_after_first',
'cohort_5_lapsed_after_first_long',
'cohort_5_lapsed_after_second',
'cohort_5_lapsed_key_multi_year',
'cohort_5_lapsed_long',
'cohort_5_lapsed_long_key_multi_year',
'cohort_5_lapsed_long_multi_year',
'cohort_5_lapsed_multi_year',
'cohort_5_multi_year',
'cohort_5_recovered_key_multi_year',
'cohort_5_recovered_multi_year',
'cohort_5_second_with_gap',
'cohort_5_second_year',
'cohort_6_first_gift',
'cohort_6_key_multi_year',
'cohort_6_lapsed_after_first',
'cohort_6_lapsed_after_first_long',
'cohort_6_lapsed_after_second',
'cohort_6_lapsed_key_multi_year',
'cohort_6_lapsed_long',
'cohort_6_lapsed_long_key_multi_year',
'cohort_6_lapsed_long_multi_year',
'cohort_6_multi_year'
]

donations_cohort_with_cycle.isnull().sum().sum()
#df.isnull().sum().sum()

# checked for nan values--one column "cohort_6_multi_year" has 545726 nan values
donations_cohort_with_cycle.isna().any()[lambda x: x]

#replace nans in cohort_6_multi_year with zeros
donations_cohort_with_cycle['cohort_6_multi_year'] = donations_cohort_with_cycle['cohort_6_multi_year'].fillna(0)

#create new dataframe with all "amount" columns to convert from cents to dollars
amount_columns = donations_cohort_with_cycle[[
'lapsed_amount',
'recovered_amount',
"amount_2013",
"amount_2014",
"amount_2015",
"amount_2016",
"amount_2017",
"lapsed_amount_2013",
"lapsed_amount_2014",
"lapsed_amount_2015",
"lapsed_amount_2016",
"lapsed_amount_2017",
"upgraded_amount_2013",
"upgraded_amount_2014",
"upgraded_amount_2015",
"upgraded_amount_2016",
"upgraded_amount_2017",
"downgraded_amount_2013",
"downgraded_amount_2014",
"downgraded_amount_2015",
"downgraded_amount_2016",
"downgraded_amount_2017",
'amount_yr0',
'amount_yr1',
'amount_yr2',
'amount_yr3',
'amount_yr4',
'lapsed_amount_yr0',
'lapsed_amount_yr1',
'lapsed_amount_yr2',
'lapsed_amount_yr3',
'lapsed_amount_yr4',
'upgraded_amount_yr0',
'upgraded_amount_yr1',
'upgraded_amount_yr2',
'upgraded_amount_yr3',
'upgraded_amount_yr4',
'downgraded_amount_yr0',
'downgraded_amount_yr1',
'downgraded_amount_yr2',
'downgraded_amount_yr3',
'downgraded_amount_yr4',
"amount_01_2013",
"amount_02_2013",
"amount_03_2013",
"amount_04_2013",
"amount_05_2013",
"amount_06_2013",
"amount_07_2013",
"amount_08_2013",
"amount_09_2013",
"amount_10_2013",
"amount_11_2013",
"amount_12_2013",
"amount_01_2014",
"amount_02_2014",
"amount_03_2014",
"amount_04_2014",
"amount_05_2014",
"amount_06_2014",
"amount_07_2014",
"amount_08_2014",
"amount_09_2014",
"amount_10_2014",
"amount_11_2014",
"amount_12_2014",
"amount_01_2015",
"amount_02_2015",
"amount_03_2015",
"amount_04_2015",
"amount_05_2015",
"amount_06_2015",
"amount_07_2015",
"amount_08_2015",
"amount_09_2015",
"amount_10_2015",
"amount_11_2015",
"amount_12_2015",
"amount_01_2016",
"amount_02_2016",
"amount_03_2016",
"amount_04_2016",
"amount_05_2016",
"amount_06_2016",
"amount_07_2016",
"amount_08_2016",
"amount_09_2016",
"amount_10_2016",
"amount_11_2016",
"amount_12_2016",
"amount_01_2017",
"amount_02_2017",
"amount_03_2017",
"amount_04_2017",
"amount_05_2017",
"amount_06_2017",
"amount_07_2017",
"amount_08_2017",
"amount_09_2017",
"amount_10_2017",
"amount_11_2017",
"amount_12_2017"]].copy()

#create new dataframe with non-amount columns
non_amount_columns = donations_cohort_with_cycle[[
'id',
'first_gift_year',
"lapsed_count",
"recovered_count",
"max_consec_giving_years",
"total_giving_years",
"count_2013",
"count_2014",
"count_2015",
"count_2016",
"count_2017",
'count_yr0',
'count_yr1',
'count_yr2',
'count_yr3',
'count_yr4',
'initial',
'first_gift',
'second_year',
'second_with_gap',
'multi_year',
'key_multi_year',
'recovered_multi_year',
'inconsistent_multi_year ',
'recovered_key_multi_year',
'lapsed_after_first',
'lapsed_after_first_long',
'lapsed_after_second',
'lapsed_multi_year',
'lapsed_long_multi_year',
'lapsed_key_multi_year',
'lapsed_long',
'cohort_0_first_gift',
'cohort_1_lapsed_after_first',
'cohort_1_second_year',
'cohort_2_lapsed_after_first_long',
'cohort_2_lapsed_after_second',
'cohort_2_multi_year',
'cohort_2_second_with_gap',
'cohort_3_first_gift',
'cohort_3_inconsistent_multi_year',
'cohort_3_key_multi_year',
'cohort_3_lapsed_after_second',
'cohort_3_lapsed_long',
'cohort_3_lapsed_multi_year',
'cohort_4_first_gift',
'cohort_4_inconsistent_multi_year',
'cohort_4_key_multi_year',
'cohort_4_lapsed_after_first',
'cohort_4_lapsed_key_multi_year',
'cohort_4_lapsed_long',
'cohort_4_lapsed_long_multi_year',
'cohort_4_lapsed_multi_year',
'cohort_4_multi_year',
'cohort_4_recovered_multi_year',
'cohort_4_second_year',
'cohort_5_first_gift',
'cohort_5_inconsistent_multi_year',
'cohort_5_key_multi_year',
'cohort_5_lapsed_after_first',
'cohort_5_lapsed_after_first_long',
'cohort_5_lapsed_after_second',
'cohort_5_lapsed_key_multi_year',
'cohort_5_lapsed_long',
'cohort_5_lapsed_long_key_multi_year',
'cohort_5_lapsed_long_multi_year',
'cohort_5_lapsed_multi_year',
'cohort_5_multi_year',
'cohort_5_recovered_key_multi_year',
'cohort_5_recovered_multi_year',
'cohort_5_second_with_gap',
'cohort_5_second_year',
'cohort_6_first_gift',
'cohort_6_key_multi_year',
'cohort_6_lapsed_after_first',
'cohort_6_lapsed_after_first_long',
'cohort_6_lapsed_after_second',
'cohort_6_lapsed_key_multi_year',
'cohort_6_lapsed_long',
'cohort_6_lapsed_long_key_multi_year',
'cohort_6_lapsed_long_multi_year',
'cohort_6_multi_year']].copy()

#convert cents to dollars in all "amount" columns
amount_columns=amount_columns/100

#merge amount and non_amount dataframes back into one
donations=non_amount_columns.merge(amount_columns, left_index=True, right_index=True)

#create new column "amount_total", which equals a donor's total donation amount 2013-2017
donations['amount_total']=donations['amount_2013'] + donations['amount_2014'] + donations['amount_2015'] + donations['amount_2016'] + donations['amount_2017']

#create new column "count_total", which equals a donor's total number of donations 2013-2017
donations['count_total']=donations['count_2013'] + donations['count_2014'] + donations['count_2015'] + donations['count_2016'] + donations['count_2017']

#giving patterns
conditions=[(donations[('amount_yr0')] >0) & (donations[('amount_yr1')] ==0) & (donations[('amount_yr2')]==0) & (donations[('amount_yr3')]==0) & (donations[('amount_yr4')]== 0) ,
(donations[('amount_yr0')] >0) & (donations[('amount_yr1')] >0) & (donations[('amount_yr2')]==0) & (donations[('amount_yr3')]==0) & (donations[('amount_yr4')]== 0) ,
(donations[('amount_yr0')] >0) & (donations[('amount_yr1')] ==0) & (donations[('amount_yr2')]>0) & (donations[('amount_yr3')]==0) & (donations[('amount_yr4')]== 0) ,
(donations[('amount_yr0')] >0) & (donations[('amount_yr1')] ==0) & (donations[('amount_yr2')]==0) & (donations[('amount_yr3')]>0) & (donations[('amount_yr4')]== 0) ,
(donations[('amount_yr0')] >0) & (donations[('amount_yr1')] ==0) & (donations[('amount_yr2')]==0) & (donations[('amount_yr3')]==0) & (donations[('amount_yr4')]> 0) ,
(donations[('amount_yr0')] >0) & (donations[('amount_yr1')] >0) & (donations[('amount_yr2')]>0) & (donations[('amount_yr3')]==0) & (donations[('amount_yr4')]== 0) ,
(donations[('amount_yr0')] >0) & (donations[('amount_yr1')] >0) & (donations[('amount_yr2')]==0) & (donations[('amount_yr3')]>0) & (donations[('amount_yr4')]== 0) ,
(donations[('amount_yr0')] >0) & (donations[('amount_yr1')] >0) & (donations[('amount_yr2')]==0) & (donations[('amount_yr3')]==0) & (donations[('amount_yr4')]> 0) ,
(donations[('amount_yr0')] >0) & (donations[('amount_yr1')] ==0) & (donations[('amount_yr2')]>0) & (donations[('amount_yr3')]>0) & (donations[('amount_yr4')]== 0) ,
(donations[('amount_yr0')] >0) & (donations[('amount_yr1')] ==0) & (donations[('amount_yr2')]>0) & (donations[('amount_yr3')]==0) & (donations[('amount_yr4')]>0) ,
(donations[('amount_yr0')] >0) & (donations[('amount_yr1')] ==0) & (donations[('amount_yr2')]==0) & (donations[('amount_yr3')]>0) & (donations[('amount_yr4')]> 0) ,
(donations[('amount_yr0')] >0) & (donations[('amount_yr1')] >0) & (donations[('amount_yr2')]>0) & (donations[('amount_yr3')]>0) & (donations[('amount_yr4')]== 0) ,
(donations[('amount_yr0')] >0) & (donations[('amount_yr1')] >0) & (donations[('amount_yr2')]>0) & (donations[('amount_yr3')]==0) & (donations[('amount_yr4')]>0) ,
(donations[('amount_yr0')] >0) & (donations[('amount_yr1')] >0) & (donations[('amount_yr2')]==0) & (donations[('amount_yr3')]>0) & (donations[('amount_yr4')]>0) ,
(donations[('amount_yr0')] >0) & (donations[('amount_yr1')] ==0) & (donations[('amount_yr2')]>0) & (donations[('amount_yr3')]>0) & (donations[('amount_yr4')]>0) ,
(donations[('amount_yr0')] >0) & (donations[('amount_yr1')] >0) & (donations[('amount_yr2')]>0) & (donations[('amount_yr3')]>0) & (donations[('amount_yr4')]>0)]


choices = ['0,',
'0,1,',
'0,2,',
'0,3,',
'0,4,',
'0,1,2,',
'0,1,3,',
'0,1,4,',
'0,2,3,',
'0,2,4,',
'0,3,4',
'0,1,2,3,',
'0,1,2,4,',
'0,1,3,4,',
'0,2,3,4,',
'0,1,2,3,4,']

donations.loc[:,'cohort_gift_pattern'] = np.select(conditions, choices, default='none')

conditions=[(donations[('amount_yr1')] >0)]
choices=['T']
donations.loc[:,'cohort_yr1_TF_short'] = np.select(conditions, choices, default='F')

conditions=[(donations[('amount_yr2')] >0)]
choices=['T']
donations.loc[:,'cohort_yr2_TF_short'] = np.select(conditions, choices, default='F')

conditions=[(donations[('amount_yr3')] >0)]
choices=['T']
donations.loc[:,'cohort_yr3_TF_short'] = np.select(conditions, choices, default='F')

conditions=[(donations[('amount_yr4')] >0)]
choices=['T']
donations.loc[:,'cohort_yr4_TF_short'] = np.select(conditions, choices, default='F')

conditions=[(donations[('amount_yr0')] >0)]
choices=['T']
donations.loc[:,'cohort_yr0_TF'] = np.select(conditions, choices, default='none')

conditions=[(donations[('amount_yr0')] >0) & (donations['amount_yr1'] > 0),
(donations[('amount_yr0')] >0) & (donations['amount_yr1'] == 0)]
choices=['TT','TF']
donations.loc[:,'cohort_yr1_TF'] = np.select(conditions, choices, default='none')

conditions=[(donations[('amount_yr0')] > 0) & (donations['amount_yr1'] > 0) & (donations['amount_yr2'] > 0),
(donations[('amount_yr0')] >0) & (donations['amount_yr1'] == 0) & (donations['amount_yr2'] > 0),
(donations[('amount_yr0')] >0) & (donations['amount_yr1'] > 0) & (donations['amount_yr2'] == 0),
(donations[('amount_yr0')] > 0) & (donations['amount_yr1'] == 0) & (donations['amount_yr2'] == 0)]
choices=['TTT','TFT','TTF','TFF']
donations.loc[:,'cohort_yr2_TF'] = np.select(conditions, choices, default='none')

conditions=[(donations[('amount_yr0')] > 0) & (donations['amount_yr1'] > 0) & (donations['amount_yr2'] > 0) & (donations['amount_yr3'] > 0),
(donations[('amount_yr0')] > 0) & (donations['amount_yr1'] > 0) & (donations['amount_yr2'] > 0) & (donations['amount_yr3'] == 0),
(donations[('amount_yr0')] >0) & (donations['amount_yr1'] == 0) & (donations['amount_yr2'] > 0) & (donations['amount_yr3'] > 0),
(donations[('amount_yr0')] >0) & (donations['amount_yr1'] == 0) & (donations['amount_yr2'] > 0) & (donations['amount_yr3'] == 0),
(donations[('amount_yr0')] > 0) & (donations['amount_yr1'] > 0) & (donations['amount_yr2'] == 0) & (donations['amount_yr3'] > 0),
(donations[('amount_yr0')] >0) & (donations['amount_yr1'] > 0) & (donations['amount_yr2'] == 0) & (donations['amount_yr3'] == 0),
(donations[('amount_yr0')] >0) & (donations['amount_yr1'] == 0) & (donations['amount_yr2'] == 0) & (donations['amount_yr3'] > 0),
(donations[('amount_yr0')] > 0) & (donations['amount_yr1'] == 0) & (donations['amount_yr2'] == 0) & (donations['amount_yr3']== 0)]

choices=['TTTT','TTTF','TFTT','TFTF','TTFT','TTFF','TFFT','TFFF']
donations.loc[:,'cohort_yr3_TF'] = np.select(conditions, choices, default='none')

conditions=[
(donations['amount_yr0']> 0) & (donations['amount_yr1']> 0) & (donations['amount_yr2']> 0) & (donations['amount_yr3']> 0) & (donations['amount_yr4']> 0),
(donations['amount_yr0']> 0) & (donations['amount_yr1']> 0) & (donations['amount_yr2']> 0) & (donations['amount_yr3']> 0) & (donations['amount_yr4']==0),
(donations['amount_yr0']> 0) & (donations['amount_yr1']> 0) & (donations['amount_yr2']> 0) & (donations['amount_yr3']==0) & (donations['amount_yr4']> 0),
(donations['amount_yr0']> 0) & (donations['amount_yr1']> 0) & (donations['amount_yr2']> 0) & (donations['amount_yr3']==0) & (donations['amount_yr4']==0),
(donations['amount_yr0']> 0) & (donations['amount_yr1']==0) & (donations['amount_yr2']> 0) & (donations['amount_yr3']> 0) & (donations['amount_yr4']> 0),
(donations['amount_yr0']> 0) & (donations['amount_yr1']==0) & (donations['amount_yr2']> 0) & (donations['amount_yr3']> 0) & (donations['amount_yr4']==0),
(donations['amount_yr0']> 0) & (donations['amount_yr1']==0) & (donations['amount_yr2']> 0) & (donations['amount_yr3']==0) & (donations['amount_yr4']> 0),
(donations['amount_yr0']> 0) & (donations['amount_yr1']==0) & (donations['amount_yr2']> 0) & (donations['amount_yr3']==0) & (donations['amount_yr4']==0),
(donations['amount_yr0']> 0) & (donations['amount_yr1']> 0) & (donations['amount_yr2']==0) & (donations['amount_yr3']> 0) & (donations['amount_yr4']> 0),
(donations['amount_yr0']> 0) & (donations['amount_yr1']> 0) & (donations['amount_yr2']==0) & (donations['amount_yr3']> 0) & (donations['amount_yr4']==0),
(donations['amount_yr0']> 0) & (donations['amount_yr1']> 0) & (donations['amount_yr2']==0) & (donations['amount_yr3']==0) & (donations['amount_yr4']> 0),
(donations['amount_yr0']> 0) & (donations['amount_yr1']> 0) & (donations['amount_yr2']==0) & (donations['amount_yr3']==0) & (donations['amount_yr4']==0),
(donations['amount_yr0']> 0) & (donations['amount_yr1']==0) & (donations['amount_yr2']==0) & (donations['amount_yr3']> 0) & (donations['amount_yr4']> 0),
(donations['amount_yr0']> 0) & (donations['amount_yr1']==0) & (donations['amount_yr2']==0) & (donations['amount_yr3']> 0) & (donations['amount_yr4']==0),
(donations['amount_yr0']> 0) & (donations['amount_yr1']==0) & (donations['amount_yr2']==0) & (donations['amount_yr3']==0) & (donations['amount_yr4']> 0),
(donations['amount_yr0']> 0) & (donations['amount_yr1']==0) & (donations['amount_yr2']==0) & (donations['amount_yr3']==0) & (donations['amount_yr4']==0)]

choices=['TTTTT', 'TTTTF', 'TTTFT', 'TTTFF', 'TFTTT', 'TFTTF', 'TFTFT', 'TFTFF', 'TTFTT', 'TTFTF', 'TTFFT', 'TTFFF', 'TFFTT', 'TFFTF', 'TFFFT', 'TFFFF']

donations.loc[:,'cohort_yr4_TF'] = np.select(conditions, choices, default='none')

donations.cohort_yr0_TF.value_counts()

#yr0_behavior
donations.loc[:,'behavior_yr0'] = np.where(donations['cohort_yr0_TF'] == "T", "new", "none")

#yr1_behavior
conditions = [donations.cohort_yr1_TF.isin(['TT']),
              donations.cohort_yr1_TF.isin(['TF'])]
            
choices = ['retain', 'lapse']

donations.loc[:,'behavior_yr1'] = np.select(conditions, choices, default='none')

#yr2_behavior
conditions = [donations.cohort_yr2_TF.isin(['TTT', 'TFT']),
              donations.cohort_yr2_TF.isin(['TFF', 'TTF'])]
            
choices = ['retain', 'lapse']

donations.loc[:,'behavior_yr2'] = np.select(conditions, choices, default='none')

#yr3_behavior
conditions = [donations.cohort_yr3_TF.isin(['TTTT', 'TFTT','TFFT', 'TTFT']),
              donations.cohort_yr3_TF.isin(['TFFF', 'TTFF', 'TFTF', 'TTTF'])]
            
choices = ['retain', 'lapse']

donations.loc[:,'behavior_yr3'] = np.select(conditions, choices, default='none')

#yr4_behavior
conditions = [donations.cohort_yr4_TF.isin(['TTTTT', 'TFTFT', 'TFTTT','TTFTT','TTTFT','TTFFT','TFFTT','TFFFT']),
              donations.cohort_yr4_TF.isin(['TTTTF', 'TFTFF', 'TFTTF','TTFTF','TTTFF','TTFFF','TFFTF','TFFFF'])]

            
choices = ['retain', 'lapse']

donations.loc[:,'behavior_yr4'] = np.select(conditions, choices, default='none')

donations.cohort_yr2_TF.value_counts()

#yr1_behavior2
conditions = [(donations['behavior_yr1'] == 'retain'),
              (donations['behavior_yr1'] == 'lapse')]
            
choices = ['second_gift', 'first_lapse']

donations.loc[:,'behavior2_yr1'] = np.select(conditions, choices, default='none')

#yr1_behavior3
conditions = [(donations['behavior_yr1'] == 'retain') & (donations['amount_yr1'] > donations['amount_yr0']),
              (donations['behavior_yr1'] == 'retain') & (donations['amount_yr1'] == donations['amount_yr0']),
              (donations['behavior_yr1'] == 'retain') & (donations['amount_yr1'] < donations['amount_yr0'])]
            
choices = ['upgrade', 'maintain', 'downgrade']

donations.loc[:,'behavior3_yr1'] = np.select(conditions, choices, default='n/a')

donations.query('amount_yr1 > 0').groupby('behavior3_yr1').count()['id'].reset_index()

#yr2_behavior2
conditions = [(donations['behavior_yr2'] == 'retain') & (donations['behavior_yr1'] == 'retain'),
              (donations['behavior_yr2'] == 'retain') & (donations['behavior_yr1'] == 'lapse'),
              (donations['behavior_yr2'] == 'lapse') & (donations['behavior_yr1'] == 'lapse'),
              (donations['behavior_yr2'] == 'lapse') & (donations['behavior_yr1'] == 'retain')]
            
choices = ['third_gift', 'recover (second_gift)', 'second_lapse', 'first_lapse']

donations.loc[:,'behavior2_yr2'] = np.select(conditions, choices, default='none')

#yr2_behavior3
conditions = [(donations['behavior_yr2'] == 'retain') & (donations['behavior_yr1'] == 'retain') & (donations['amount_yr2'] > donations['amount_yr1']),
              (donations['behavior_yr2'] == 'retain') & (donations['behavior_yr1'] == 'retain') & (donations['amount_yr2'] == donations['amount_yr1']),
              (donations['behavior_yr2'] == 'retain') & (donations['behavior_yr1'] == 'retain') & (donations['amount_yr2'] < donations['amount_yr1']),
              (donations['behavior_yr2'] == 'retain') & (donations['behavior_yr1'] == 'lapse') & (donations['amount_yr2'] > donations['amount_yr0']),
              (donations['behavior_yr2'] == 'retain') & (donations['behavior_yr1'] == 'lapse') & (donations['amount_yr2'] == donations['amount_yr0']),
              (donations['behavior_yr2'] == 'retain') & (donations['behavior_yr1'] == 'lapse') & (donations['amount_yr2'] < donations['amount_yr0'])]
            
choices = ['upgrade', 'maintain', 'downgrade', 'upgrade', 'maintain', 'downgrade']

donations.loc[:,'behavior3_yr2'] = np.select(conditions, choices, default='n/a')

donations.query('behavior_yr3 == "retain"').cohort_yr3_TF.value_counts()

print(donations.cohort_yr1_TF.value_counts())
print(donations.cohort_yr2_TF.value_counts())
print(donations.cohort_yr3_TF.value_counts())
print(donations.cohort_yr4_TF.value_counts())

#yr2_behavior3
conditions = [donations['cohort_yr3_TF'] == 'TFFF',
              donations['cohort_yr3_TF'] == 'TTFF',
              donations['cohort_yr3_TF'] == 'TFTF',
              donations['cohort_yr3_TF'] == 'TTTF',
              donations['cohort_yr3_TF'] == 'TTTT',
              donations['cohort_yr3_TF'] == 'TFFT',
              donations['cohort_yr3_TF'] == 'TTFT',
              donations['cohort_yr3_TF'] == 'TFTT']
              
choices = ['third_lapse', 'second_lapse', 'second_lapse', 'first_lapse', 'fourth_gift', 'recover(second_gift)', 'recover(third_gift)', 'third_gift']

donations.loc[:,'behavior3_yr3'] = np.select(conditions, choices, default='none')

#yr2_behavior3
conditions = [(donations['behavior_yr3'] == 'retain') & (donations['behavior_yr2'] == 'retain') & (donations['behavior_yr1'] == 'retain') & (donations['amount_yr3'] > donations['amount_yr2']),
              (donations['behavior_yr3'] == 'retain') & (donations['behavior_yr2'] == 'retain') & (donations['behavior_yr1'] == 'retain') & (donations['amount_yr3'] == donations['amount_yr2']),              
              (donations['behavior_yr3'] == 'retain') & (donations['behavior_yr2'] == 'retain') & (donations['behavior_yr1'] == 'retain') & (donations['amount_yr3'] < donations['amount_yr2']),
              (donations['behavior_yr3'] == 'retain') & (donations['behavior_yr2'] == 'lapse') & (donations['behavior_yr1'] == 'lapse') & (donations['amount_yr3'] > donations['amount_yr0']),
              (donations['behavior_yr3'] == 'retain') & (donations['behavior_yr2'] == 'lapse') & (donations['behavior_yr1'] == 'lapse') & (donations['amount_yr3'] == donations['amount_yr0']),              
              (donations['behavior_yr3'] == 'retain') & (donations['behavior_yr2'] == 'lapse') & (donations['behavior_yr1'] == 'lapse') & (donations['amount_yr3'] < donations['amount_yr0']),
              (donations['behavior_yr3'] == 'retain') & (donations['behavior_yr2'] == 'lapse') & (donations['behavior_yr1'] == 'retain') & (donations['amount_yr3'] > donations['amount_yr1']),
              (donations['behavior_yr3'] == 'retain') & (donations['behavior_yr2'] == 'lapse') & (donations['behavior_yr1'] == 'retain') & (donations['amount_yr3'] == donations['amount_yr1']),              
              (donations['behavior_yr3'] == 'retain') & (donations['behavior_yr2'] == 'lapse') & (donations['behavior_yr1'] == 'retain') & (donations['amount_yr3'] < donations['amount_yr1']),
              (donations['behavior_yr3'] == 'retain') & (donations['behavior_yr2'] == 'retain') & (donations['behavior_yr1'] == 'lapse') & (donations['amount_yr3'] > donations['amount_yr2']),
              (donations['behavior_yr3'] == 'retain') & (donations['behavior_yr2'] == 'retain') & (donations['behavior_yr1'] == 'lapse') & (donations['amount_yr3'] == donations['amount_yr2']),              
              (donations['behavior_yr3'] == 'retain') & (donations['behavior_yr2'] == 'retain') & (donations['behavior_yr1'] == 'lapse') & (donations['amount_yr3'] < donations['amount_yr2']),
              (donations['behavior_yr3'] == 'lapse') & (donations['behavior_yr2'] == 'retain') & (donations['behavior_yr1'] == 'retain'),
              (donations['behavior_yr3'] == 'lapse') & (donations['behavior_yr2'] == 'retain') & (donations['behavior_yr1'] == 'lapse'),
              (donations['behavior_yr3'] == 'lapse') & (donations['behavior_yr2'] == 'lapse') & (donations['behavior_yr1'] == 'retain'),
              (donations['behavior_yr3'] == 'lapse') & (donations['behavior_yr2'] == 'lapse') & (donations['behavior_yr1'] == 'lapse')]
              
choices = ['upgrade', 'maintain', 'downgrade', 'upgrade', 'maintain', 'downgrade','upgrade', 'maintain', 'downgrade', 'upgrade', 'maintain', 'downgrade', 'first_lapse', 'second_lapse', 'second_lapse', 'third_lapse']

donations.loc[:,'behavior3_yr3'] = np.select(conditions, choices, default='none')

### REMINDER: NEED TO ADD behavior2_yr4 and behavior3_yr4

#donations.drop('cohort_TF_pattern', axis=1, inplace=True)

donations['first_gift_year_year'] = pd.DatetimeIndex(donations['first_gift_year']).year

conditions = [
    (donations['first_gift_year_year'] == 2013),
    (donations['first_gift_year_year'] == 2014),
    (donations['first_gift_year_year'] == 2015),
    (donations['first_gift_year_year'] == 2016),
    (donations['first_gift_year_year'] == 2017),
    (donations['first_gift_year_year'] == 2018)    
    ]

# create a list of the values we want to assign for each condition
values = ['cohort13', 'cohort14', 'cohort15', 'cohort16', 'cohort17', 'cohort18']

# create a new column and use np.select to assign values to it using our lists as arguments
donations['cohort'] = np.select(conditions, values)

#Create three donation count groups:  (i) never more than 1 donation / year,  (ii) at most, 2-52 donations/year, & (iii) at least once made greater than 52 donations/year.
conditions=[(donations['count_yr0'] <= 1) & (donations['count_yr1'] <= 1) & (donations['count_yr2'] <= 1) & (donations['count_yr3'] <= 1) & (donations['count_yr4'] <= 1),
            ((donations['count_yr0'] <= 52) & (donations['count_yr1'] <= 52) & (donations['count_yr2'] <= 52) & (donations['count_yr3'] <= 52) & (donations['count_yr4'] <= 52) & ((donations['count_yr0'] > 1) | (donations['count_yr1'] > 1) | (donations['count_yr2'] > 1) | (donations['count_yr3'] > 1) | (donations['count_yr4'] > 1))),
            (donations['count_yr0'] > 52) | (donations['count_yr1'] > 52) | (donations['count_yr2'] > 52) | (donations['count_yr3'] > 52) | (donations['count_yr4'] > 52)]

choices=['1x', '2-52x', '>52x']
donations.loc[:, 'donation_counts']=np.select(conditions, choices, default='none')

#create column grouping by yr0 donation amounts
conditions=[(donations['amount_yr0'] <= 25),
            (donations['amount_yr0'] > 25) & (donations['amount_yr0'] <= 50),
            (donations['amount_yr0'] > 50) & (donations['amount_yr0'] <= 100),
            (donations['amount_yr0'] > 100) & (donations['amount_yr0'] <= 1000),
            (donations['amount_yr0'] > 1000)]
    
choices=['under25', '25to50', '50to100', '100to1000', 'over1000']

donations.loc[:,'yr0_amount_groups']=np.select(conditions, choices, default='none')

donations.to_pickle('donations_cohort_with_cycle.pkl')

# # donors by # contributions
# l = 12
# for n in range(1,l+1,1):
#     if n < l + 1:
#         count=count_df.loc[count_df['count_col'] == n]
#         print(n, f'/',
#               "{:,}".format(count.count_col.count()), f'/',
#               "{:.2%}".format(count.count_col.count()/count_df.count_col.count()), f'/',
#              "${:,.0f}".format(count.amount_col.sum()), f'/',
#              "{:.2%}".format(count.amount_col.sum()/count_df.amount_col.sum()))

#donations=pd.read_pickle('donations.pkl')
