==> notebooks/15a_handling_donation_counts_DONE.ipynb <==
import numpy as np
import pandas as pd

import plotly.express as px
import plotly.graph_objects as go
import plotly.io as pio
from plotly.subplots import make_subplots
import plotly.figure_factory as ff

from ipywidgets import widgets

#import re
# import dash
# import dash_core_components as dcc
# import dash_html_components as html

import seaborn as sns
import matplotlib.pyplot as plt
sns.set_style("white")
%matplotlib inline

donations=pd.read_pickle('donations_cohort_with_cycle.pkl')

## Assessing donation counts.

#### 1. Create three donation count groups:  (i) never more than 1 donation / year,  (ii) at most, 2-52 donations/year, & (iii) at least once made greater than 52 donations/year.

print(f'number donors donating 1x/yr: ', '{:,.0f}'.format(donations.query('count_yr0 <= 1 & count_yr1 <= 1 & count_yr2 <= 1 & count_yr3 <= 1 & count_yr4 <= 1').count()['id']))
print(f'percentage of total donors: ', '{:.2%}'.format(donations.query('count_yr0 <= 1 & count_yr1 <= 1 & count_yr2 <= 1 & count_yr3 <= 1 & count_yr4 <= 1').count()['id']/donations.count()['id']))

print(f'number donors donating 2-52x/yr (at least one year), but never more than 52x/year: ', 
      '{:,.0f}'.format(donations.query('count_yr0 <= 52 & count_yr1 <= 52 & count_yr2 <= 52 & count_yr3 <= 52 & count_yr4 <= 52 & (count_yr0 > 1 | count_yr1 > 1 | count_yr2 > 1 | count_yr3 > 1 | count_yr4 > 1)').count()['id']))
print(f'percentage of total donors: ', '{:.2%}'.format((donations.query('count_yr0 <= 52 & count_yr1 <= 52 & count_yr2 <= 52 & count_yr3 <= 52 & count_yr4 <= 52 & (count_yr0 > 1 | count_yr1 > 1 | count_yr2 > 1 | count_yr3 > 1 | count_yr4 > 1)').count()['id'])/donations.count()['id']))

print(f'number donors donating > 52 times/yr (at least one year): ', 
      '{:,.0f}'.format(donations.query('count_yr0 > 52 | count_yr1 > 52 | count_yr2 > 52 | count_yr3 > 52 | count_yr4 > 52').count()['id']))
print(f'percentage of total donors: ', 
      '{:.2%}'.format(donations.query('count_yr0 > 52 | count_yr1 > 52 | count_yr2 > 52 | count_yr3 > 52 | count_yr4 > 52').count()['id']/donations.count()['id']))

conditions=[(donations['count_yr0'] <= 1) & (donations['count_yr1'] <= 1) & (donations['count_yr2'] <= 1) & (donations['count_yr3'] <= 1) & (donations['count_yr4'] <= 1),
            ((donations['count_yr0'] <= 52) & (donations['count_yr1'] <= 52) & (donations['count_yr2'] <= 52) & (donations['count_yr3'] <= 52) & (donations['count_yr4'] <= 52) & ((donations['count_yr0'] > 1) | (donations['count_yr1'] > 1) | (donations['count_yr2'] > 1) | (donations['count_yr3'] > 1) | (donations['count_yr4'] > 1))),
            (donations['count_yr0'] > 52) | (donations['count_yr1'] > 52) | (donations['count_yr2'] > 52) | (donations['count_yr3'] > 52) | (donations['count_yr4'] > 52)]

choices=['1x', '2-52x', '>52x']
donations.loc[:, 'donation_counts']=np.select(conditions, choices, default='none')

#add this column into notebook 10 and create pickle.
