==> notebooks/16_creating_database_for_sql.ipynb <==
import numpy as np
import pandas as pd

import plotly.express as px
import plotly.graph_objects as go
import plotly.io as pio
from plotly.subplots import make_subplots
import plotly.figure_factory as ff

from ipywidgets import widgets

#import re
# import dash
# import dash_core_components as dcc
# import dash_html_components as html

import seaborn as sns
import matplotlib.pyplot as plt
sns.set_style("white")
%matplotlib inline

import mysql.connector

from sqlalchemy import create_engine

donations=pd.read_pickle('donations_cohort_with_cycle.pkl')

donations=donations.rename(columns={'inconsistent_multi_year ':'inconsistent_multi_year'})

db = mysql.connector.connect(
    host = "localhost",
    user = "root",
    passwd = "200broaD!"
)

cursor = db.cursor()

## executing the statement using 'execute()' method
cursor.execute("SHOW DATABASES")

## 'fetchall()' method fetches all the rows from the last executed statement
databases = cursor.fetchall() ## it returns a list of all databases present

## printing the list of databases
print(databases)

## showing one by one database
for database in databases:
    print(database)

engine = create_engine("mysql://root:[password]@localhost:3306/donations"
                   #   fast_executemany=True
                      )
con = engine.connect()
sample.to_sql(name='sample',con=engine, if_exists='append')
con.close()
